fix(mobile): update code struct
